requirements.txt
python-telegram-bot==20.3
pypdf
reportlab
PyMuPDF
